refactor: Flatten consensus::verify::* to consensus::*
